headers reorg for comms deployment
